feat : made custom hook to get notes array
